requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-ollama>=0.1.0
python-dotenv>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
beautifulsoup4>=4.12.0
